made order history screen functional

+++ lib/features/dashboard/domain/entities/cart_entity.dart
@@ -3,7 +3,7 @@ import 'product_entity.dart';
 class CartItem {
   final Product product;
   final int quantity;
-  final int? size;
+  final double? size;
   final bool? isHot;
   final int? sugar;
   final String? milk;
